loadbalancer: Move the legacy Backend and SVC to legacy.go
Move the Backend and SVC types used by the old control-plane to
legacy.go in preparation for renaming them.
